Extract logic for spawning worker into its own module
